refactor(chat): simplify attachment upload route structure
Simplifies attachment upload endpoint from `/api/v/[agentId]/[sessionId]/attachments`
to `/api/attachments/upload`, removing unnecessary path parameters.

Changes:
- Create new route at /api/attachments/upload
- Remove agentId validation (always "c010" - hardcoded everywhere)
- Remove sessionId parameter (never used - storage uses userId from auth)
- Update useAttachmentUpload hook to use new endpoint
- Remove agentId/sessionId parameters from hook interface
- Delete old route file

Rationale:
- sessionId was in route params but never referenced in handler logic
- agentId was only used to validate against constant "c010"
- User scoping already handled via Clerk auth (userId)
- Cleaner, more RESTful API design
